+++ CHANGES.rst
@@ -10,6 +10,7 @@ Changelog (nionui)
 - Add support for directly getting last or current keyboard modifiers.
 - Only change canvas item focus when no key modifiers are enabled.
 - Add menu methods for adding only if enabled and checking if enabled.
+- Expand action architecture to support parameters and results.
 
 0.4.5 (2021-02-02)
 ------------------